auth_system/requirements.txt
Django==4.2.10
djangorestframework==3.14.0
psycopg2==2.9.9
python-dotenv==1.0.0
